New error GHCBackendError.NotAHaskellType instead of GenericError
