+++ src/full/Agda/Compiler/MAlonzo/Misc.hs
@@ -29,6 +29,7 @@ import Agda.TypeChecking.Monad
 
 import Agda.Syntax.Common.Pretty
 
+import Agda.Utils.CallStack ( HasCallStack )
 import Agda.Utils.Impossible
 
 --------------------------------------------------
@@ -145,6 +146,9 @@ runHsCompileT' t e s = (flip runStateT s) . (flip runReaderT e) $ t
 
 runHsCompileT :: HsCompileT m a -> GHCModuleEnv -> m (a, HsCompileState)
 runHsCompileT t e = runHsCompileT' t e mempty
+
+ghcBackendError :: (HasCallStack, MonadTCError m) => GHCBackendError -> m a
+ghcBackendError = typeError . GHCBackendError
 
 --------------------------------------------------
 -- utilities for haskell names
