+++ test/Compiler/simple/Erased-cubical-FFI.out
@@ -1,7 +1,7 @@
 COMPILE_FAILED
 
 ret > ExitFailure 42
-out > /Erased-cubical-FFI.agda:13,1-35: error: [GenericDocError]
+out > /Erased-cubical-FFI.agda:13,1-35: error: [GHCBackend.NotAHaskellType]
 out > The type Not-compiled → Bool cannot be translated to a
 out > corresponding Haskell type, because it contains a name that is not
 out > compiled (Not-compiled).
